+++ src/utilities.js
@@ -221,6 +221,8 @@ exports.checkArgs = async function (args, fromWhere) {
 	let goodValues;
 	let badKeys = [];
 	let badValues = [];
+  
+  const simpleKeys = ["restrictFind", "reveal", "runWhen", "runPostCommands"];
 
 	if (fromWhere === "findBinding" || fromWhere === "findSetting") {
 		goodKeys = findCommands.getKeys();     // an array
@@ -260,17 +262,8 @@ exports.checkArgs = async function (args, fromWhere) {
           if (!goodValues[key].includes(value)) badValues.push({ [key]: value });
         }));
       }
-      // combine these using an array
-      else if (key === 'restrictFind') {  // combine these next 3?
-        if (!goodValues[key].includes(args[key])) badValues.push({ [key]: args[key] });
-      }
-      else if (key === 'reveal') {    // TODO check this
-        if (!goodValues[key].includes(args[key])) badValues.push({ [key]: args[key] });
-      }
-      else if (key === 'runWhen') {
-        if (!goodValues[key].includes(args[key])) badValues.push({ [key]: args[key] });
-      }
-      else if (key === 'runPostCommands') {
+      
+      else if (simpleKeys.includes(key)) {
         if (!goodValues[key].includes(args[key])) badValues.push({ [key]: args[key] });
       }