+++ src/variables.js
@@ -8,7 +8,6 @@ exports.getExtensionDefinedVariables = function() {
 }
 
 
-
 /**
  * @returns {Array} - all the available path variables
  */